central_service/requirements.txt
Flask==2.2.2
python-dotenv
pymongo[srv]
pika
google-cloud-storage==2.5.*
metrohash
